Stage 2

+++ src/main/java/main/api/response/PostResponse.java
@@ -10,7 +10,7 @@ public class PostResponse {
     private List<Post> posts;
 
     public PostResponse() {
-        count = (long) 0;
+        count = 0L;
         posts = new ArrayList<>();
     }
